feat(summary): applyTransaction

+++ src/main/java/com/alvayonara/finguardriskservice/transaction/TransactionService.java
@@ -20,7 +20,7 @@ public class TransactionService {
                                     saved.getId(),
                                     saved.getUserId(),
                                     saved.getType(),
-                                    saved.getAmount(),
+                                    saved.getAmount().doubleValue(),
                                     saved.getOccurredAt().toString()
                             );
                     transactionEventPublisher.publishTransactionCreated(event);
